lambdas/python/requirements.txt
boto3>=1.24.2
python-magic>=0.4.27
Pylibmagic>=0.2.2